requirements.txt
fastapi
uvicorn
requests
python-dotenv
radon
transformers
torch
scikit-learn
shap
pandas
numpy
joblib